refactor: moved translationSteps mapping and userAgentInfo to own hook
