+++ src/mobile/feature-identification/scan/components/PermissionDeniedInstructions.tsx
@@ -1,35 +1,16 @@
+import { BoldFormattedText } from "@/mobile/lib/ui/uitpas/BoldFormattedText";
 import { useTranslation } from "@/shared/lib/utils/hooks";
+import { useUserAgentInfo } from "@/shared/lib/utils/hooks/useUserAgentInfo";
 import { Box, List, ListItem, ListItemText, Typography } from "@mui/material";
-import { ReactNode, useMemo } from "react";
-import uaParser from "ua-parser-js";
+import { useMemo } from "react";
+import { useMapTranslationsToSteps } from "./permissionDeniedInstructionsMapper";
 
-type Step = {
-  key: string;
-  subSteps: string[];
-};
 
 export const PermissionDeniedInstructions = () => {
   const { t } = useTranslation();
+  const mapTranslationsToSteps = useMapTranslationsToSteps()
 
-  const formatBoldText = (text: string): ReactNode[] => {
-    if (!text.includes("**")) return [text];
-
-    const parts = text.split(/(\*\*.*?\*\*)/g);
-    return parts.map((part, index) => {
-      if (part.startsWith("**") && part.endsWith("**")) {
-        const boldText = part.slice(2, -2);
-        return <strong key={index}>{boldText}</strong>;
-      }
-      return part;
-    });
-  };
-
-  const userAgentInfo = useMemo(() => {
-    if (typeof window !== "undefined") {
-      return uaParser(navigator.userAgent);
-    }
-    return { browser: { name: undefined }, os: { name: undefined } };
-  }, []);
+  const userAgentInfo = useUserAgentInfo()
 
   const variant = useMemo(() => {
     if (!userAgentInfo.browser.name) return "chrome";
@@ -57,49 +38,19 @@ export const PermissionDeniedInstructions = () => {
     return "chrome";
   }, [userAgentInfo]);
 
-  const getSteps = (variant: string): Step[] => {
-    const steps: Step[] = [];
-    let stepIndex = 1;
-    let stepKey = `identification.mobile.scan.permissionDenied.steps.${variant}.step${stepIndex}`;
-
-    if (variant !== "chrome" && t(stepKey, { defaultValue: "" }) === "") {
-      variant = "chrome";
-      stepKey = `identification.mobile.scan.permissionDenied.steps.${variant}.step${stepIndex}`;
-    }
-
-    while (t(stepKey, { defaultValue: "" }) !== "") {
-      const step: Step = {
-        key: stepKey,
-        subSteps: [],
-      };
-
-      let subStepIndex = 1;
-      let subStepKey = `identification.mobile.scan.permissionDenied.steps.${variant}.step${stepIndex}.${subStepIndex}`;
-
-      while (t(subStepKey, { defaultValue: "" }) !== "") {
-        step.subSteps.push(subStepKey);
-        subStepIndex++;
-        subStepKey = `identification.mobile.scan.permissionDenied.steps.${variant}.step${stepIndex}.${subStepIndex}`;
-      }
-
-      steps.push(step);
-      stepIndex++;
-      stepKey = `identification.mobile.scan.permissionDenied.steps.${variant}.step${stepIndex}`;
-    }
-
-    return steps;
-  };
+  const steps = mapTranslationsToSteps(variant)
 
   return (
     <Box sx={{ width: "100%", maxWidth: 600 }}>
       <Typography variant="body1" sx={{ mb: 2 }}>
-        {formatBoldText(
+        <BoldFormattedText text={
           t("identification.mobile.scan.permissionDenied.description")
-        )}
+        }/>
+        
       </Typography>
 
       <List sx={{ p: 0, pl: 2 }}>
-        {getSteps(variant).map((step, index) => (
+        {steps.map((step, index) => (
           <Box key={step.key}>
             <ListItem
               sx={{
@@ -109,7 +60,7 @@ export const PermissionDeniedInstructions = () => {
                 pl: 1,
               }}
             >
-              <ListItemText primary={formatBoldText(t(step.key))} />
+              <ListItemText primary={<BoldFormattedText text={t(step.key)} />}/>
             </ListItem>
 
             {step.subSteps.length > 0 && (
@@ -139,10 +90,10 @@ export const PermissionDeniedInstructions = () => {
                       >
                         {subStepNumber}
                       </Box>
-                      <ListItemText
-                        primary={formatBoldText(t(subStepKey))}
-                        sx={{ m: 0 }}
-                      />
+                        <ListItemText
+                          primary={<BoldFormattedText text={t(subStepKey)} />}
+                          sx={{ m: 0 }}
+                        />
                     </ListItem>
                   );
                 })}
